comp-practice/lab_9/requirements.txt
Pillow
exif
